+++ CHANGELOG.md
@@ -58,6 +58,7 @@ RELEASING:
 - Exclude snakeyaml [#1418](https://github.com/GIScience/openrouteservice/pull/1418)
 - Upgrade geotools from 28.2 to 29.0 [#1422](https://github.com/GIScience/openrouteservice/pull/1422)
 - upgrade graphhopper version to v4.6 [#1427](https://github.com/GIScience/openrouteservice/pull/1427)
+- Map matching of traffic data ([#1430](https://github.com/GIScience/openrouteservice/pull/1430))
 
 ## [7.0.1] - 2023-03-08
 ### Fixed